Add a type cast of UINT32 to insure that EBC compiler does not report warning of "Integer conversion resulted in a change of sign" when building with -Za option (accept only ANSI/ISO standard C code). 1 << 31 is consider to be a sign change by compiler.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5325 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Universal/Network/IScsiDxe/IScsiProto.h
@@ -31,13 +31,13 @@ Abstract:
 #define ISCSI_SEQ_EQ(s1, s2)  ((s1) == (s2))
 #define ISCSI_SEQ_LT(s1, s2) \
     ( \
-      (((INT32) (s1) < (INT32) (s2)) && (s2 - s1) < (1 << 31)) || \
-      (((INT32) (s1) > (INT32) (s2)) && (s1 - s2) > (1 << 31)) \
+      (((INT32) (s1) < (INT32) (s2)) && (s2 - s1) < ((UINT32) 1 << 31)) || \
+      (((INT32) (s1) > (INT32) (s2)) && (s1 - s2) > ((UINT32) 1 << 31)) \
     )
 #define ISCSI_SEQ_GT(s1, s2) \
     ( \
-      (((INT32) (s1) < (INT32) (s2)) && (s2 - s1) > (1 << 31)) || \
-      (((INT32) (s1) > (INT32) (s2)) && (s1 - s2) < (1 << 31)) \
+      (((INT32) (s1) < (INT32) (s2)) && (s2 - s1) > ((UINT32) 1 << 31)) || \
+      (((INT32) (s1) > (INT32) (s2)) && (s1 - s2) < ((UINT32) 1 << 31)) \
     )
 
 #define ISCSI_WELL_KNOWN_PORT                   3260
